requirements.txt
discord.py
python-dotenv
aiohttp
streamlit
pandas
plotly

solana
solders
websockets
base58
